Added per-account and per-mediaType cache ttl configuration - refactored
