+++ src/test/java/org/prebid/server/auction/ExchangeServiceTest.java
@@ -38,10 +38,10 @@ import org.prebid.server.bidder.Usersyncer;
 import org.prebid.server.bidder.model.BidderBid;
 import org.prebid.server.bidder.model.BidderError;
 import org.prebid.server.bidder.model.BidderSeatBid;
-import org.prebid.server.cache.CacheBid;
-import org.prebid.server.cache.CacheIdInfo;
 import org.prebid.server.cache.CacheService;
 import org.prebid.server.cache.account.AccountCacheService;
+import org.prebid.server.cache.model.CacheBid;
+import org.prebid.server.cache.model.CacheIdInfo;
 import org.prebid.server.cache.model.CacheTtl;
 import org.prebid.server.cookie.UidsCookie;
 import org.prebid.server.currency.CurrencyConversionService;
